refactor: move window_iterator to a independent file
