+++ src/storage/mem_table.cc
@@ -25,6 +25,7 @@
 #include "common/timer.h"
 #include "gflags/gflags.h"
 #include "storage/record.h"
+#include "storage/window_iterator.h"
 
 DECLARE_uint32(skiplist_max_height);
 DECLARE_uint32(skiplist_max_height);
@@ -774,113 +775,6 @@ bool MemTable::BulkLoad(const std::vector<DataBlock*>& data_blocks,
     }
 
     return true;
-}
-
-MemTableKeyIterator::MemTableKeyIterator(Segment** segments, uint32_t seg_cnt, ::openmldb::storage::TTLType ttl_type,
-                                         uint64_t expire_time, uint64_t expire_cnt, uint32_t ts_index)
-    : segments_(segments),
-      seg_cnt_(seg_cnt),
-      seg_idx_(0),
-      pk_it_(NULL),
-      it_(NULL),
-      ttl_type_(ttl_type),
-      expire_time_(expire_time),
-      expire_cnt_(expire_cnt),
-      ticket_(),
-      ts_idx_(0) {
-    uint32_t idx = 0;
-    if (segments_[0]->GetTsIdx(ts_index, idx) == 0) {
-        ts_idx_ = idx;
-    }
-}
-
-MemTableKeyIterator::~MemTableKeyIterator() {
-    if (pk_it_ != NULL) delete pk_it_;
-}
-
-void MemTableKeyIterator::SeekToFirst() {
-    ticket_.Pop();
-    if (pk_it_ != NULL) {
-        delete pk_it_;
-        pk_it_ = NULL;
-    }
-    for (seg_idx_ = 0; seg_idx_ < seg_cnt_; seg_idx_++) {
-        pk_it_ = segments_[seg_idx_]->GetKeyEntries()->NewIterator();
-        pk_it_->SeekToFirst();
-        if (pk_it_->Valid()) return;
-        delete pk_it_;
-        pk_it_ = NULL;
-    }
-}
-
-void MemTableKeyIterator::Seek(const std::string& key) {
-    if (pk_it_ != NULL) {
-        delete pk_it_;
-        pk_it_ = NULL;
-    }
-    ticket_.Pop();
-    if (seg_cnt_ > 1) {
-        seg_idx_ = ::openmldb::base::hash(key.c_str(), key.length(), SEED) % seg_cnt_;
-    }
-    Slice spk(key);
-    pk_it_ = segments_[seg_idx_]->GetKeyEntries()->NewIterator();
-    pk_it_->Seek(spk);
-    if (!pk_it_->Valid()) {
-        NextPK();
-    }
-}
-
-bool MemTableKeyIterator::Valid() {
-    return pk_it_ != NULL && pk_it_->Valid();
-}
-
-void MemTableKeyIterator::Next() { NextPK(); }
-
-::hybridse::vm::RowIterator* MemTableKeyIterator::GetRawValue() {
-    TimeEntries::Iterator* it = NULL;
-    if (segments_[seg_idx_]->GetTsCnt() > 1) {
-        KeyEntry* entry = ((KeyEntry**)pk_it_->GetValue())[ts_idx_];  // NOLINT
-        it = entry->entries.NewIterator();
-        ticket_.Push(entry);
-    } else {
-        it = ((KeyEntry*)pk_it_->GetValue())  // NOLINT
-                 ->entries.NewIterator();
-        ticket_.Push((KeyEntry*)pk_it_->GetValue());  // NOLINT
-    }
-    it->SeekToFirst();
-    return new MemTableWindowIterator(it, ttl_type_, expire_time_, expire_cnt_);
-}
-
-std::unique_ptr<::hybridse::vm::RowIterator> MemTableKeyIterator::GetValue() {
-    return std::unique_ptr<::hybridse::vm::RowIterator>(GetRawValue());
-}
-
-const hybridse::codec::Row MemTableKeyIterator::GetKey() {
-    hybridse::codec::Row row(
-        ::hybridse::base::RefCountedSlice::Create(pk_it_->GetKey().data(), pk_it_->GetKey().size()));
-    return row;
-}
-
-void MemTableKeyIterator::NextPK() {
-    do {
-        ticket_.Pop();
-        if (pk_it_->Valid()) {
-            pk_it_->Next();
-        }
-        if (!pk_it_->Valid()) {
-            delete pk_it_;
-            pk_it_ = NULL;
-            seg_idx_++;
-            if (seg_idx_ < seg_cnt_) {
-                pk_it_ = segments_[seg_idx_]->GetKeyEntries()->NewIterator();
-                pk_it_->SeekToFirst();
-                if (!pk_it_->Valid()) {
-                    continue;
-                }
-            }
-        }
-        break;
-    } while (true);
 }
 
 MemTableTraverseIterator::MemTableTraverseIterator(Segment** segments, uint32_t seg_cnt,
